Aggregate raster values over polygon areas

Takes the mean of raster values in an area, ignoring any missing values.

Usage

augment_epi_units_with_raster(
  epi_units,
  raster,
  risk_name = "raster_risk",
  aggregate_fun = c("mean", "max", "min", "sum")
)

Arguments

epi_units

polygon data, usually epidemiologcal units (EU).

raster

raster data to take the aggregate of each polygon in p.

risk_name

name of new column containing aggregated raster values

aggregate_fun

function to use to aggregate raster values over epi units area, default is mean.

Value

returns the polygon data (sf) with a new column containing the aggregated values from the raster data for each polygon.