Calculate weighted border risk

Calculates the risk of introduction for each epidemiological unit based on the risk of emission of neighbouring countries. The score is weighted based on the length of shared borders.

Usage

calc_border_risk(epi_units, shared_borders, emission_risk)

Arguments

epi_units

epidemiological units dataset

shared_borders

shared borders dataset as outputted by calc_border_lengths()

emission_risk

emission risk dataset,

Value

An sf object containing the border risk score associated with each epidemiological unit. The table is containing the following columns:

eu_id: ID of the epidemiological unit. This will be used as a join key to the epi_units table and risk_table.

eu_name: Name of the epidemiological unit. This comes from the epi_units table.

borders: Total length of shared borders with neighbouring countries in km.

border_risk: The risk of introduction based on the emission risk of neighbouring countries, its values are in the range [0, 12].

geometry: The geometry column containing the shared border, a MULTIPOLYGON.

This dataset also has a number of attributes, the first is a dataset, the other 3 are used in other functions from riskintroanalysis to make passing dataset metadata between functions more user-friendly.

borders: an sf object containing shared borders between EUs and bordering countries. The table is containing the following columns:

eu_id: ID of the epidemiological unit.

bc_id: ISO3 of the bordering country.

border_length: Length of the shared border, use units::units to get the units. It is expected in km.

geometry: The geometry column containing the shared border, a MULTILINESTRING or a LINESTRING.

table_name = "border_risk": the name of the table in the riskintroanalysis data model. This is used by other functions in riskintroanalysis, such as plot_risk.

risk_col = "border_risk": the name of the risk score column in the dataset. This is used by other functions in riskintroanalysis, such as plot_risk.

scale = c(0,12): the scale of risk_col used for other functions such as rescale_risk_score and plot_risk.

Examples

# \donttest{

library(sf)
library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

tunisia_raw <- sf::read_sf(system.file(
  package = "riskintrodata",
  "samples", "tunisia", "epi_units", "tunisia_adm2_raw.gpkg"
))

# Apply mapping to prepare colnames and validate dataset
tunisia <- apply_mapping(
  tunisia_raw,
  mapping = mapping_epi_units(
    eu_name = "NAME_2",
    geometry = "geom"
  ),
  validate = TRUE
)
#> ✔ All data in "epi_units" valided.

tun_neighbours <- riskintrodata::neighbours_table |>
  filter(country_id == "TUN")

bordering_countries <- riskintrodata::world_sf |>
  filter(iso3 %in% tun_neighbours$neighbour_id) |>
  filter(iso3 %in% "LBY")

# Run function to get shared borders
shared_borders <- calc_border_lengths(
  epi_units = tunisia,
  eu_id_col = "eu_id",
  bordering_countries = bordering_countries,
  bc_id_col = "iso3"
)

wahis_factors <- riskintrodata::get_wahis_erf(
  disease = "Anthrax",
  species = "Cattle",
  animal_category = "Domestic"
)
#> ✔ All data in "emission_risk_factors" valided.
#> ✔ WAHIS emission risk factors dataset has 65 entries for `disease = Anthrax`, `species = Cattle`, and `animal_category = Domestic`.

emission_risk_table <- calc_emission_risk(wahis_factors)

ri_border <- calc_border_risk(
  epi_units = tunisia,
  shared_borders = shared_borders,
  emission_risk = emission_risk_table
)
ri_border
#> Simple feature collection with 268 features and 4 fields
#> Geometry type: GEOMETRY
#> Dimension:     XY
#> Bounding box:  xmin: 7.530076 ymin: 30.23681 xmax: 11.59826 ymax: 37.55986
#> Geodetic CRS:  WGS 84
#> # A tibble: 268 × 5
#>    eu_id    eu_name            borders border_risk                      geometry
#>    <chr>    <chr>                <dbl>       <dbl>                <GEOMETRY [°]>
#>  1 eu-00001 Ariana Médina           NA           0 POLYGON ((10.13862 36.89432,…
#>  2 eu-00002 Ettadhamen              NA           0 POLYGON ((10.05623 36.84281,…
#>  3 eu-00003 Kalaat El Andalous      NA           0 MULTIPOLYGON (((10.13862 36.…
#>  4 eu-00004 Mnihla                  NA           0 POLYGON ((10.13109 36.88332,…
#>  5 eu-00005 Raoued                  NA           0 POLYGON ((10.17025 36.88844,…
#>  6 eu-00006 Sebkhet Ariana          NA           0 POLYGON ((10.27149 36.88916,…
#>  7 eu-00007 Sidi Thabet             NA           0 POLYGON ((10.00166 36.99815,…
#>  8 eu-00008 Soukra                  NA           0 POLYGON ((10.20232 36.85612,…
#>  9 eu-00009 Amdoun                  NA           0 POLYGON ((9.136977 36.87276,…
#> 10 eu-00010 Béja Nord               NA           0 POLYGON ((9.091034 36.70194,…
#> # ℹ 258 more rows

border_lines <- extract_border(ri_border)
border_lines
#> Simple feature collection with 3 features and 9 fields
#> Geometry type: GEOMETRY
#> Dimension:     XY
#> Bounding box:  xmin: 9.557298 ymin: 30.23681 xmax: 11.59826 ymax: 33.15514
#> Geodetic CRS:  WGS 84
#> # A tibble: 3 × 10
#>   eu_id bc_id border_length                  geometry weight country data_source
#>   <chr> <chr>         <dbl>            <GEOMETRY [°]>  <dbl> <chr>   <chr>      
#> 1 eu-0… LBY            130. LINESTRING (11.56026 33.…      1 NA      NA         
#> 2 eu-0… LBY            143. MULTILINESTRING ((10.615…      1 NA      NA         
#> 3 eu-0… LBY            176. LINESTRING (10.13751 31.…      1 NA      NA         
#> # ℹ 3 more variables: emission_risk <dbl>, border_risk <dbl>,
#> #   border_label <list>
# }