Build the weighted emission risk table

This function calculates the overall emission risk score (emission_risk) for a set of countries based on four domains (each one counts for a certain number of points towards the risk of introduction out of 12 by default) :

Epidemiological status (3/12): Time since the last outbreak (sc_epistatus).

Surveillance measures (2/12): Effectiveness of implemented surveillance strategies (sc_survmeasures).

Control measures (3/12): Effectiveness of disease control measures (sc_control).

Animal commerce movements (4/12): Risk from commerce and movement of animals (sc_commerce).

The emission risk is the sum of each of the above.

Each of the scores is calculated from the emission risk factors and some are weighted based on the emission risk factor weights (weights parameter).

sc_epistatus: The epidemiological status score is based on the time since the last outbreak and accounds for 3 out of 12 of the final emission risk score. If the disease has not been detected in the last 5 years (\(x > 5\)), the score is 0. If the disease is currently present (\(x = 0\)), the score is 3. An exponential decay model smooths the scoring over time: $$S = 3 \times \exp\left(-x \frac{\log(2)}{5}\right)$$

sc_survmeasures: Surveillance measures are scored based on the absence of effective measures and accounts for 2 out of 12 og the final emission risk score. The following risk factors contribute to this score:

Active surveillance

Passive surveillance

Risk-based surveillance

Mandatory reporting

sc_control: Control measures are scored similarly, and account for 3 out of 12 of the emission risk score. The following risk factors contribute to this score:

Border control

Culling at outbreak sites

Culling around outbreak sites

Movement zoning and restrictions

Ring vaccination around outbreak sites

sc_commerce: The risk score for animal commerce movements can take values of 0, 1, 3, or 4. It is the sum of the following:

0 means there is no legal or illegal trade

legal trade present adds 1 to this score

illegal trade present adds 3 to this score

The overall emission risk score is the sum of each weighted risk factor, the final emission risk score is in the range of (0, 12].

Usage

calc_emission_risk(
  emission_risk_factors,
  weights = get_erf_weights(),
  keep_scores = TRUE
)

Arguments

emission_risk_factors

A data frame containing risk factor data. Generally, this data will come from riskintrodata::get_wahis_erf(). The dataset should be validated and have table_name attribute equal to "emission_risk_factors".

weights

A named list of weights corresponding to the following columns in emission_risk_factors (and their default weights):

disease_notification (0.25)

targeted_surveillance (0.5)

general_surveillance (0.5)

screening (0.75)

precautions_at_the_borders (1)

slaughter (0.5)

selective_killing_and_disposal (0.5)

zoning (0.75)

official_vaccination (0.25)

The sum of the weights should add up to exactly 5, as these factors correspond to the weights contributing to sc_survmeasures and sc_control.

keep_scores

whether to keep or drop sc_* columns, emission_risk column is always kept.

Value

A tibble containing the following columns:

iso3 identifis the country

country country name (from emission risk factors dataset)

disease disease being studied (from emission risk factors dataset)

animal_category animal_category being studied (from emission risk factors dataset)

species species being studied (from emission risk factors dataset)

data_source data source for emission risk factors (from emission risk factors dataset)

sc_survmeasures as detailed above.

sc_control as detailed above.

sc_commerce as detailed above.

sc_epistatus as detailed above.

emission_risk as detailed above.

This dataset also has a number of attributes that are used in other functions from riskintroanalysis to make passing dataset metadata between functions more user-friendly. Used mainly in used by plot_risk() and rescale_risk_scores().

table_name = "emission_risk_scores"

risk_col = "emission_risk"

scale = c(0, 12)

table_validated = TRUE

Examples

library(riskintrodata)
library(riskintroanalysis)

wahis_erf <- get_wahis_erf(
  disease = "Avian infectious laryngotracheitis",
  animal_category = "Domestic",
  species = "Birds"
)
#> ✔ All data in "emission_risk_factors" valided.
#> ✔ WAHIS emission risk factors dataset has 62 entries for `disease = Avian infectious laryngotracheitis`, `species = Birds`, and `animal_category = Domestic`.

emission_risk_table <- calc_emission_risk(
  emission_risk_factors = wahis_erf
)