Road access introduction risk analysis

Calculates the risk of introduction for an animal disease using an accessibility to roads index.

Usage

calc_road_access_risk(
  epi_units,
  road_access_raster,
  aggregate_fun = c("mean", "max", "min", "sum")
)

Arguments

epi_units

epidemiological units dataset

road_access_raster

road access raster data contains road accessibility index. Can be aquired using riskintrodata::download_road_access_raster() and terra::rast().

aggregate_fun

function to use to aggregate raster values over epi units area, default is mean.

Value

list with class ri_analysis that contains:

ri: an sf dataset containing the road accessibility risk summarised over each epidemiological unit. Contains the following columns:

eu_id: unique identifier for each epidemiological unit

eu_name: name provided in epi_units dataset.

road_access_risk: road access risk of inroduction score (unscaled)

geometry: geometry for each epidemiological unit of type POLYGON or MULTIPOLYGON.

This data can be extracted from the ri_analysis object with extract_raster()

raster: a SpatRaster containing the cropped raster that covers only the epidemiological units areas.

See also

Examples

if (FALSE) { # \dontrun{
library(riskintroanalysis)
library(riskintrodata)
library(dplyr)
library(terra)
library(sf)
library(ggplot2)
# Example with raw sf files, previously downloaded with geodata::gadm()
tunisia_raw <- read_sf(system.file(
  package = "riskintrodata",
  "samples", "tunisia", "epi_units", "tunisia_adm2_raw.gpkg"
))

# Apply mapping to prepare colnames and validate dataset
tunisia <- apply_mapping(
  tunisia_raw,
  mapping = mapping_epi_units(
    eu_name = "NAME_2",
    geometry = "geom"
  ),
  validate = TRUE
)

road_raster_fp <- download_road_access_raster()
road_raster <- rast(road_raster_fp)

road_access_risk <- calc_road_access_risk(
  epi_units = tunisia,
  road_access_raster = road_raster,
  aggregate_fun = "mean"
)

# plotting the raster data
raster <- extract_raster(road_access_risk)
plot(raster)

# plotting the risk of introduction dataset
ggplot(data=road_access_risk, aes(fill = road_access_risk)) + geom_sf()

} # }