Extract analysis subcomponents

Helper functions to access subcomponents of riskintro_analysis objets.

Usage

extract_point_risk(x)

extract_flow_risk(x)

extract_raster(x)

extract_border(x)

Arguments

x

analysis object containing multiple steps/components.

Value

Returns the analysis step depending on the fuction :

extract_point_risk(x) returns the point risk, usually the first step in entry point introduction risk analysis. Contains weighted and aggregated emission risk data for each entry points.

extract_flow_risk(x) returns the flow risk, usually the first step in animal mobility introduction risk analysis. Contains weighted and aggregated emission risk data for each flow.

extract_raster(x) returns the raster data cropped to the epidemiological units areas.

extract_border(x) returns the border risk data for cropped to the epidemiological units areas.